+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-v2.5.4
+v2.5.4 (2024-01-24)
 ------
 * Switching on a string field now acts upon a stable local variable copy.
 * Added a method to insert code after a label.